third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFESpecularLightingElement-svgdom-suraceScale-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'surfaceScale' property of the SVGFESpecularLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS specularElement.surfaceScale.baseVal is 1
PASS specularElement.surfaceScale.baseVal is 10
PASS successfullyParsed is true

TEST COMPLETE
